Implemented the Bitcoin block height randomization of top 20% weighted peers (max 5)
We might want to add more sources for Bitcoin blocks, just so we're not
vulnerable to the 2 sites listed here.

Added EOSIOP2P address to the Discovery file, which is going to be
required for the network pre-mesh.

Simplified hooks, now that there's just "exec".

Added notes to `ipfs` run so you can run multiple instances on the same machine
(for simulating many testnets)