Updated to dawn-v4.2.0. ChainID is retrieved through `get_info` now.
Renamed `privkeys.keys` to `seed_network.keys`.

Keep `eosio` on schedule, to complete the last transactions.

Renamed some ops.

TODO: get to the bottom of `delegatebw`'s `transfer`. Also, `eosio` ends
up with mucho money.  But he delegated so much bw.. not where where all
that EOS goes.. we need to figure out how to compute that.

After this launch, 118.0000 EOS are in `eosio.ram`.. might be more
with the whole snapshot.  Difficult to know in advance how much that RAM
costs.  No special case for `eosio` in the contract.

First thing BPs could do is burn the remainers of `eosio`.

Updated boot sequence, and ops, and my_discovery_file and published
all the v4.2.0 contents.