Propose edits
Line 5 - remove curly braces around `regaccount` as the curly braces are used to print a bit of information pulled from elsewhere. `regaccount` is the just the name of this contract.
Line 5 - also the Ethereum address and balance are both values that will be pulled.
Line 7 - Adding in the chain ID to show that the intent of the contract is to create an account here, and not on some other network. Give the full intention.
Line 11 - Added DR forum